feat: create and use useAuthorization
