+++ webapp/src/components/Modals/SellModal/SellModal.module.css
@@ -6,6 +6,14 @@
   margin: 5%;
 }
 
+.loaderContainer {
+  align-items: center;
+  display: flex;
+  height: 200px;
+  justify-content: center;
+  width: 100%;
+}
+
 :global(.dcl.field) {
   width: 100% !important;
   align-self: center;
